- ok, this should enable Bobs portaudio fixes and ALSA depending on   --enable options.
This will have to be cleaned up a lot more to minimise #ifdef's
a2d_() is needed for portaudio which is needed by both windows and *nix
but only a small piece of start_threads.c is needed for portaudio
the ALSA part is #ifdef'd out when USE_PORTAUDIO is enabled.
audio_init.f90 and a2d.f90 now have #ifdef's on USE_PORTAUDIO
instead of Win32.

Lots to do still.



git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/trunk@73 ab8295b8-cf94-4d9e-aec4-7959e3be5d79