Added CMake option to control experimental features.
Sometimes parts of the application are  not fully ready for release at
release  time.   This  option  allows  such  features  to  be  enabled
conditionally at configure time.

The CMake option WSJT_ENABLE_EXPERIMENTAL_FEATURES  is used to control
this ability.  Normally  you will not need to change  it as it behaves
in a reasonable  way.  The option is ON in  debug build configurations
and OFF in non-debug configuratiuons.

The  option  is   made  available  in  all  source   modules  via  the
WSJT_ENABLE_EXPERIMENTAL_FEATURES macro  which is defined to  1 if the
option is ON and to 0 if it is OFF.

Since this  default behavior is  potentially dangerous because  it can
mean that release builds contain different code from debug builds; the
option may  be set to  OFF in debug  configurations in the  normal way
(cmake-gui,  edit  CMakeCache.txt, etc.)   to  achieve  a debug  build
similar to a default release build.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4182 ab8295b8-cf94-4d9e-aec4-7959e3be5d79