Use default PTT type in Hamlib when CAT is selected.
Hamlib knows  for each  rig back-end  which type  of PTT  switching is
available,  we  don't  need  to  set  it  as  the  default  is  always
correct. Setting it to "RIG" was disabling the ability to use the rear
audio  input  select version  of  CAT  PTT  on  rigs that  support  it
e.g. Kenwood TS-590s.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4224 ab8295b8-cf94-4d9e-aec4-7959e3be5d79