Improve socket write handling in HRD interface.
The TCP/IP socket write handling in teh HRD inteface was naive and did
not handle  incomplete writes t all.  This change loops until  all the
payload is written for each command sent.

Also error  handling is improved  so that a  socket in an  error state
cannot get used inadvertenly.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4263 ab8295b8-cf94-4d9e-aec4-7959e3be5d79