Add a CMake option to allow a Debug wsjtx.exe to be a console application
The normal way  to view console output from a  Windows GUI application
(i.e. with a  WinMain entry point) is to attach  a debugger which will
then receive the application console output in its own console.

If it is  required to build wsjtx.exe as a  console application it may
be  done  by  setting  the new  WSJT_CREATE_WINMAIN  CMake  option  to
OFF. This option  is only available for Debug  configuration builds to
avoid  accidental  shipping of  non-Debug  configuration  builds as  a
console application.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@4555 ab8295b8-cf94-4d9e-aec4-7959e3be5d79