WSJT-X User Guide location variable
In order to support the WSJT-X  super build in LOCAL mode the location
of the User Guide needs to be specified.

A CMake cache  variable WSJT_MANUAL_DIRECTORY has been  added to allow
the URI of the User Guide directory to be specified at configure time.
The default value remains as the project web site.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx-1.4@4565 ab8295b8-cf94-4d9e-aec4-7959e3be5d79