Reduced FFTW patience to 1 for debug builds for fastest startup
Don't make  proper Windows  applications in  debug builds.  This means
that  the  debug output  now  goes  to  the  console and  the  program
(wsjtx.exe) no longer detaches from the console.

If you  require a proper Windows  executable, for instance if  you are
getting  the  debug  output  via  the normal  method  of  attaching  a
debugger,  you can  still do  so  by setting  the CMake  configuration
option WSJT_CREATE_WINMAIN to ON for your debug build configuration on
Windows.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5249 ab8295b8-cf94-4d9e-aec4-7959e3be5d79