Use absolute path for Python interpreter to avoid picking up v3
Because asciidoc  doesn't work with Python  3 we must not  allow it to
run via  a shebang  or Windows  file association,  instead we  now use
find_package()  on Python  and  filter the  version.  This means  that
CMAKE_PREFIX_PATH   can  be   used  to   select  the   correct  Python
installation if there are more than one and one is v3 or newer.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@5321 ab8295b8-cf94-4d9e-aec4-7959e3be5d79