Handle disappearing displays correctly
Qt  makes sure  a widget's  window is  displayed on  a display  if the
display configuration  changes while  it is  running.  If  the display
configuration changes while  the application is not  running then only
QWidget::saveGeometry()  and  QWidget::restoreGeometry()  will  ensure
that  the  widget's window  is  correctly  displayed if  the  original
display is no longer present.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6257 ab8295b8-cf94-4d9e-aec4-7959e3be5d79