Reverting r6531 and replacing with correct fix
The WSPR  message is unpacked  into a symbol  per byte array  which is
processed  by  encoding a  whole  number  of  bytes of  message,  this
requires the output  array to be bigger than the  number of symbols to
accommodate  extra values  that are  not part  of the  message due  to
rounding up to whole bytes. I.e.  176 (11*8*2) elements to contain the
162 symbols.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6532 ab8295b8-cf94-4d9e-aec4-7959e3be5d79