Add support for HRD control of rigs with an HRD Data drop down
Rigs like  the Icom IC-7100  have a  Data drop down  on the HRD  UI to
switch  the rig  in and  out  and query  the data  modes. This  change
supports that by using the On and Off selections to set and query data
mode for suitable modes (USB/LSB/FM as DIG_U/DIG_L/DIG_FM).

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6599 ab8295b8-cf94-4d9e-aec4-7959e3be5d79