Ignore connection refused errors on UDP sockets
In Qt 5.6 the optional connection refused error from a UDP socket that
is written  to where no server  is listening is signalled.  This error
only happens when  the o/s can tell  that no one is  listening i.e. on
the local host loopback address. We  can safely ignore these errors as
we do not care if anyone is listening or not.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6701 ab8295b8-cf94-4d9e-aec4-7959e3be5d79