Create the splash screen before assigning the application name
Even  though the  splash screen  does  not have  a title  it seems  to
interfere  with JTAlert  which  enumerates window  titles to  identify
WSJT-X instances. I'm not sure why  this is necessary, but if it helps
it should be benign.

Ensure that the  splash screen is hidden before any  message boxes are
shown, this is essential  at startup if there is a  stale lock file or
rig control error.

Be far less aggressive about bringing  the splash screen to the top of
the application window  stack. This may cause the splash  screen to be
obscured on some platforms.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7025 ab8295b8-cf94-4d9e-aec4-7959e3be5d79