Enable OpenMP for everything except on Macs
Also increase  Windows stack  size from  1Mbyte to  8Mbyte due  to the
impact  of Fortran  arrays  not being  automatically  moved to  static
storage above a certain size. This needs attention by setting the SAVE
attribute on large arrays that do not  need to be on the stack i.e. do
no need to be duplicated across OpenMP thread teams.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7122 ab8295b8-cf94-4d9e-aec4-7959e3be5d79