Refactor the MultiSettings class to work on Linux with Qt 5.5
This  change  to  using  dynamically   built  pop  up  menus  for  teh
configurations menu gets around an  issue with teh Linux Unity appmenu
not being able to handle actions used in multiple menu items.

There is  a remaining issue  with the QMenu::aboutToHide()  signal not
being fired if the Unity appmenu  is not disabled. Hopefully this will
be fixed  in later  Qt/Unity versions.  Until then  the configurations
menu cannot be used unless the appmenu_qt5 package is removed.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7260 ab8295b8-cf94-4d9e-aec4-7959e3be5d79